feat: реализована вся функциональность

+++ src/utils/Api.js
@@ -50,19 +50,34 @@ class Api {
         })
             .then(this._checkResponse)
     }
-    likeCard(cardId) {
-        return fetch(`${this.url}/${this._cardUrl}/likes/${cardId}`, {
-            method: 'PUT',
-            headers: this.headers,
-        })
-            .then(this._checkResponse)
-    }
-    deleteLikeCard(cardId) {
-        return fetch(`${this.url}/${this._cardUrl}/likes/${cardId}`, {
-            method: 'DELETE',
-            headers: this.headers,
-        })
-            .then(this._checkResponse)
+    // likeCard(cardId) {
+    //     return fetch(`${this.url}/${this._cardUrl}/likes/${cardId}`, {
+    //         method: 'PUT',
+    //         headers: this.headers,
+    //     })
+    //         .then(this._checkResponse)
+    // }
+    // deleteLikeCard(cardId) {
+    //     return fetch(`${this.url}/${this._cardUrl}/likes/${cardId}`, {
+    //         method: 'DELETE',
+    //         headers: this.headers,
+    //     })
+    //         .then(this._checkResponse)
+    // }
+    changeLikeCardStatus(cardId, isLiked) {
+        if (isLiked) {
+            return fetch(`${this.url}/${this._cardUrl}/likes/${cardId}`, {
+                method: 'PUT',
+                headers: this.headers,
+            })
+                .then(this._checkResponse)
+        } else {
+            return fetch(`${this.url}/${this._cardUrl}/likes/${cardId}`, {
+                method: 'DELETE',
+                headers: this.headers,
+            })
+                .then(this._checkResponse)
+        }
     }
     updateTask(object, url) {
         return fetch(`${this.url}/${url}`, {
